made a test page for testing connection with backend
added in a test page that makes use of two services to connect to back end.

sqlapi service
this is where we get and post the restful http requests

classmanager
this is where we create classe for the data we use and pass them into the places they will be used. we could use model for this but this way is more versatile.